import img/wiki_up and file_gallery* content from 1.9 version
As suggested here:

    http://doc.tiki.org/Upgrade#Fresh_install_recommended_

The file_gallery* stuff looks like it should be in a subdirectory
but a feature/bug of tiki put them in the root directory.

Hopefuly this can be fixed later.

There is no content to import from img/trackers and the image
gallery is stored in the database.