relocate the file gallery files to file_gallery/
The file gallery content was stored in the website root directory.
This was an accident of missing out the final slash when initially
setting it up, and the result is messy.  So I've moved the files into
file_gallery/, by physically moving the files (which start with file_gallery),
renaming to remove the file_gallery prefix, and updating the file gallery
path settings.